feat: new revamped ready handler, should be more organized and modular!
