+++ src/handler/events/readyEvent.ts
@@ -3,12 +3,9 @@ import { basename } from 'path';
 import * as Files from '../utilities/readFile';
 import type Wrapper from '../structures/wrapper';
 import type { Result } from 'ts-results';
-import { Err, Ok } from 'ts-results';
 import type { Awaitable } from 'discord.js';
 import { ApplicationCommandType, ComponentType } from 'discord.js';
 import type { CommandModule } from '../structures/module';
-import { match } from 'ts-pattern';
-import { SernError } from '../structures/errors';
 import type { DefinedCommandModule } from '../../types/handler';
 import { CommandType, PluginType, PayloadType } from '../structures/enums';
 import { errTap } from './observableHandling';
@@ -67,62 +64,62 @@ export function onReady(wrapper: Wrapper) {
             }),
         )
         .subscribe(({ mod, cmdPluginsRes }) => {
-            const loadedPluginsCorrectly = cmdPluginsRes.every(({ execute }) => execute.ok);
-            if (loadedPluginsCorrectly) {
-                const res = registerModule(mod);
-                if (res.err) {
-                    throw Error(SernError.NonValidModuleType);
-                }
-                wrapper.sernEmitter?.emit('module.register', {
-                    type: PayloadType.Success,
-                    module: mod,
-                });
-            } else {
-                wrapper.sernEmitter?.emit('module.register', {
-                    type: PayloadType.Failure,
-                    module: mod,
-                    reason: SernError.PluginFailure,
-                });
-            }
+            // const loadedPluginsCorrectly = cmdPluginsRes.every(({ execute }) => execute.ok);
+            // if (loadedPluginsCorrectly) {
+            //     const res = registerModule(mod);
+            //     if (res.err) {
+            //         throw Error(SernError.NonValidModuleType);
+            //     }
+            //     wrapper.sernEmitter?.emit('module.register', {
+            //         type: PayloadType.Success,
+            //         module: mod,
+            //     });
+            // } else {
+            //     wrapper.sernEmitter?.emit('module.register', {
+            //         type: PayloadType.Failure,
+            //         module: mod,
+            //         reason: SernError.PluginFailure,
+            //     });
+            // }
         });
 }
 
-function registerModule(mod: DefinedCommandModule): Result<void, void> {
-    const name = mod.name;
-    return match<DefinedCommandModule>(mod)
-        .with({ type: CommandType.Text }, mod => {
-            mod.alias?.forEach(a => Files.TextCommands.aliases.set(a, mod));
-            Files.TextCommands.text.set(name, mod);
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.Slash }, mod => {
-            Files.ApplicationCommands[ApplicationCommandType.ChatInput].set(name, mod);
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.Both }, mod => {
-            Files.BothCommands.set(name, mod);
-            mod.alias?.forEach(a => Files.TextCommands.aliases.set(a, mod));
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.MenuUser }, mod => {
-            Files.ApplicationCommands[ApplicationCommandType.User].set(name, mod);
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.MenuMsg }, mod => {
-            Files.ApplicationCommands[ApplicationCommandType.Message].set(name, mod);
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.Button }, mod => {
-            Files.ApplicationCommands[ComponentType.Button].set(name, mod);
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.MenuSelect }, mod => {
-            Files.MessageCompCommands[ComponentType.SelectMenu].set(name, mod);
-            return Ok.EMPTY;
-        })
-        .with({ type: CommandType.Modal }, mod => {
-            Files.ModalSubmitCommands.set(name, mod);
-            return Ok.EMPTY;
-        })
-        .otherwise(() => Err.EMPTY);
-}
+// function registerModule(mod: DefinedCommandModule): Result<void, void> {
+//     const name = mod.name;
+//     return match<DefinedCommandModule>(mod)
+//         .with({ type: CommandType.Text }, mod => {
+//             mod.alias?.forEach(a => Files.TextCommands.aliases.set(a, mod));
+//             Files.TextCommands.text.set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.Slash }, mod => {
+//             Files.ApplicationCommands[ApplicationCommandType.ChatInput].set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.Both }, mod => {
+//             Files.BothCommands.set(name, mod);
+//             mod.alias?.forEach(a => Files.TextCommands.aliases.set(a, mod));
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.MenuUser }, mod => {
+//             Files.ApplicationCommands[ApplicationCommandType.User].set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.MenuMsg }, mod => {
+//             Files.ApplicationCommands[ApplicationCommandType.Message].set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.Button }, mod => {
+//             Files.ApplicationCommands[ComponentType.Button].set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.MenuSelect }, mod => {
+//             Files.MessageCompCommands[ComponentType.SelectMenu].set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .with({ type: CommandType.Modal }, mod => {
+//             Files.ModalSubmitCommands.set(name, mod);
+//             return Ok.EMPTY;
+//         })
+//         .otherwise(() => Err.EMPTY);
+// }
