+++ src/handler/events/readyHandler.ts
@@ -1,13 +1,19 @@
 import { EventsHandler } from './eventsHandler';
 import type Wrapper from '../structures/wrapper';
-import { concatMap, fromEvent, Observable, map, take, of } from 'rxjs';
+import { concatMap, fromEvent, Observable, map, take, of, from, toArray, switchMap } from 'rxjs';
 import type { CommandModule } from '../structures/module';
 import * as Files from '../utilities/readFile';
 import { errTap } from './observableHandling';
 import type { DefinedCommandModule } from '../../types/handler';
 import { basename } from 'path';
-import { PayloadType } from '../structures/enums';
+import { CommandType, PayloadType, PluginType } from '../structures/enums';
 import { processCommandPlugins } from './userDefinedEventsHandling';
+import type { Awaitable } from 'discord.js';
+import type { Result } from 'ts-results';
+import { SernError } from '../structures/errors';
+import { match } from 'ts-pattern';
+import { Err, Ok } from 'ts-results';
+import { ApplicationCommandType, ComponentType } from 'discord.js';
 
 export default class ReadyHandler extends EventsHandler<{
     mod: DefinedCommandModule;
@@ -31,7 +37,30 @@ export default class ReadyHandler extends EventsHandler<{
             ),
         );
         this.init();
-        this.payloadSubject.pipe(concatMap(payload => this.processPlugins(payload))).subscribe();
+        this.payloadSubject
+            .pipe(
+                concatMap(payload => this.processPlugins(payload)),
+                concatMap(payload => this.resolvePlugins(payload)),
+            )
+            .subscribe(payload => {
+                const allPluginsSuccessful = payload.pluginRes.every(({ execute }) => execute.ok);
+                if (allPluginsSuccessful) {
+                    const res = registerModule(payload.mod);
+                    if (res.err) {
+                        throw Error(SernError.NonValidModuleType);
+                    }
+                    wrapper.sernEmitter?.emit('module.register', {
+                        type: PayloadType.Success,
+                        module: payload.mod,
+                    });
+                } else {
+                    wrapper.sernEmitter?.emit('module.register', {
+                        type: PayloadType.Failure,
+                        module: payload.mod,
+                        reason: SernError.PluginFailure,
+                    });
+                }
+            });
     }
     private static intoDefinedModule({ absPath, mod }: { absPath: string; mod: CommandModule }): {
         absPath: string;
@@ -45,6 +74,33 @@ export default class ReadyHandler extends EventsHandler<{
                 ...mod,
             },
         };
+    }
+
+    private resolvePlugins({
+        mod,
+        cmdPluginRes,
+    }: {
+        mod: DefinedCommandModule;
+        cmdPluginRes: {
+            name: string;
+            description: string;
+            execute: Awaitable<Result<void, void>>;
+            type: PluginType.Command;
+        }[];
+    }) {
+        if (mod.plugins.length === 0) {
+            return of({ mod, pluginRes: [] });
+        }
+        // modules with no event plugins are ignored in the previous
+        return from(cmdPluginRes).pipe(
+            switchMap(pl =>
+                from(pl.execute).pipe(
+                    map(execute => ({ ...pl, execute })),
+                    toArray(),
+                ),
+            ),
+            map(pluginRes => ({ mod, pluginRes })),
+        );
     }
 
     private processPlugins(payload: { mod: DefinedCommandModule; absPath: string }) {
@@ -61,4 +117,44 @@ export default class ReadyHandler extends EventsHandler<{
     protected setState(state: { absPath: string; mod: DefinedCommandModule }): void {
         this.payloadSubject.next(state);
     }
+}
+
+function registerModule(mod: DefinedCommandModule): Result<void, void> {
+    const name = mod.name;
+    return match<DefinedCommandModule>(mod)
+        .with({ type: CommandType.Text }, mod => {
+            mod.alias?.forEach(a => Files.TextCommands.aliases.set(a, mod));
+            Files.TextCommands.text.set(name, mod);
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.Slash }, mod => {
+            Files.ApplicationCommands[ApplicationCommandType.ChatInput].set(name, mod);
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.Both }, mod => {
+            Files.BothCommands.set(name, mod);
+            mod.alias?.forEach(a => Files.TextCommands.aliases.set(a, mod));
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.MenuUser }, mod => {
+            Files.ApplicationCommands[ApplicationCommandType.User].set(name, mod);
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.MenuMsg }, mod => {
+            Files.ApplicationCommands[ApplicationCommandType.Message].set(name, mod);
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.Button }, mod => {
+            Files.ApplicationCommands[ComponentType.Button].set(name, mod);
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.MenuSelect }, mod => {
+            Files.MessageCompCommands[ComponentType.SelectMenu].set(name, mod);
+            return Ok.EMPTY;
+        })
+        .with({ type: CommandType.Modal }, mod => {
+            Files.ModalSubmitCommands.set(name, mod);
+            return Ok.EMPTY;
+        })
+        .otherwise(() => Err.EMPTY);
 }
